requirements.txt
anyio==4.3.0
blinker==1.6.2
branca==0.6.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
exceptiongroup==1.2.0
Flask==3.0.0
flask-blueprint==1.3.0
Flask-Cors==4.0.0
folium==0.14.0
future==0.18.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.4
iso8601==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.7.2
MarkupSafe==2.1.3
numpy==1.26.0
psycopg2==2.9.9
python-telegram-bot==21.0.1
PyYAML==6.0.1
requests==2.31.0
serial==0.0.97
sniffio==1.3.1
typing_extensions==4.10.0
urllib3==2.0.6
Werkzeug==3.0.0
